Remaning pieces.
