+++ modules/geometry/misc/python/test/test_solvepnp.py
