+++ samples/cpp/tutorial_code/calib3d/real_time_pose_estimation/src/Mesh.cpp
@@ -6,7 +6,7 @@
  */
 
 #include "Mesh.h"
-#include <opencv2/geometry.hpp>
+#include <opencv2/ptcloud.hpp>
 
 // --------------------------------------------------- //
 //                   TRIANGLE CLASS                    //
